+++ collect_app/src/main/java/org/odk/collect/android/activities/FormDownloadListActivity.java
@@ -42,11 +42,11 @@ import org.odk.collect.android.formmanagement.FormDownloader;
 import org.odk.collect.android.formmanagement.FormSourceExceptionMapper;
 import org.odk.collect.android.formmanagement.ServerFormDetails;
 import org.odk.collect.android.formmanagement.ServerFormsDetailsFetcher;
+import org.odk.collect.android.formmanagement.downloaderror.FormsDownloadErrorItem;
 import org.odk.collect.android.fragments.dialogs.FormDownloadResultDialog;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.listeners.DownloadFormsTaskListener;
 import org.odk.collect.android.listeners.FormListDownloaderListener;
-import org.odk.collect.android.logic.FormDownloadErrorItem;
 import org.odk.collect.android.network.NetworkStateProvider;
 import org.odk.collect.android.openrosa.HttpCredentialsInterface;
 import org.odk.collect.android.tasks.DownloadFormListTask;
@@ -657,10 +657,10 @@ public class FormDownloadListActivity extends FormListActivity implements FormLi
 
         DialogUtils.dismissDialog(RefreshFormListDialogFragment.class, getSupportFragmentManager());
 
-        ArrayList<FormDownloadErrorItem> failures = new ArrayList<>();
+        ArrayList<FormsDownloadErrorItem> failures = new ArrayList<>();
         for (Map.Entry<ServerFormDetails, String> entry : result.entrySet()) {
             if (!entry.getValue().equals(getString(R.string.success))) {
-                failures.add(new FormDownloadErrorItem(entry.getKey().getFormName(), entry.getKey().getFormId(), entry.getKey().getFormVersion(), entry.getValue()));
+                failures.add(new FormsDownloadErrorItem(entry.getKey().getFormName(), entry.getKey().getFormId(), entry.getKey().getFormVersion(), entry.getValue()));
             }
         }
