+++ collect_app/src/main/java/org/odk/collect/android/formmanagement/downloaderror/FormsDownloadErrorActivity.kt
@@ -1,12 +1,11 @@
-package org.odk.collect.android.activities
+package org.odk.collect.android.formmanagement.downloaderror
 
 import android.os.Bundle
 import androidx.appcompat.widget.Toolbar
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import org.odk.collect.android.R
-import org.odk.collect.android.adapters.FormsDownloadErrorAdapter
-import org.odk.collect.android.logic.FormDownloadErrorItem
+import org.odk.collect.android.activities.CollectAbstractActivity
 import org.odk.collect.android.utilities.TranslationHandler
 
 class FormsDownloadErrorActivity : CollectAbstractActivity() {
@@ -14,7 +13,7 @@ class FormsDownloadErrorActivity : CollectAbstractActivity() {
         const val FAILURES = "FAILURES"
     }
 
-    private lateinit var failures: List<FormDownloadErrorItem>
+    private lateinit var failures: List<FormsDownloadErrorItem>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -23,7 +22,7 @@ class FormsDownloadErrorActivity : CollectAbstractActivity() {
         supportActionBar!!.setDisplayHomeAsUpEnabled(true)
         findViewById<Toolbar>(R.id.toolbar).setNavigationOnClickListener { finish() }
 
-        failures = intent.getSerializableExtra(FAILURES) as List<FormDownloadErrorItem>
+        failures = intent.getSerializableExtra(FAILURES) as List<FormsDownloadErrorItem>
         findViewById<RecyclerView>(R.id.errors).apply {
             adapter = FormsDownloadErrorAdapter(failures)
             layoutManager = LinearLayoutManager(context)
